1.修复了创建新笔记，主动保存笔记后再修改标题会用空白文本覆盖原始内容的问题； 2.修复文本导出时空白文本会导致异常的问题

+++ app/src/main/java/com/work/lazxy/writeaway/ui/filter/RevokeInputFilter.java
@@ -15,36 +15,43 @@ import java.util.Stack;
  */
 
 public class RevokeInputFilter implements InputFilter {
-    private Stack<Action> mBackStack; //这里的栈中计划存储三种不同的操作符，分别是添加，删除和替换
-    private final int MAX_ITEM_IN_STACK = 20;
+    private final int MAX_ITEM_IN_STACK = 100;
+    /**
+     * 这里的栈中计划存储三种不同的操作符，分别是添加，删除和替换
+     */
+    private Stack<Action> mBackStack;
 
-    public RevokeInputFilter(Stack<Action> stack){
+    public RevokeInputFilter(Stack<Action> stack) {
         mBackStack = stack;
     }
 
     @Override
     public CharSequence filter(CharSequence source, int start, int end, Spanned dest, int dstart, int dend) {
         String origin = source.toString();
-        if(origin.startsWith(Action.REVOKED_SIGN)&&origin.endsWith(Action.REVOKED_SIGN)){
+        if (origin.startsWith(Action.IGNORE_SIGN) && origin.endsWith(Action.IGNORE_SIGN)) {
+            return origin.substring(Action.IGNORE_SIGN.length(), origin.length() - Action.IGNORE_SIGN.length());
+        }
+        if (origin.startsWith(Action.REVOKED_SIGN) && origin.endsWith(Action.REVOKED_SIGN)) {
             //如果这时是一个回退操作
-            return origin.substring(3,origin.length()-3);
-        }else{
+            return origin.substring(Action.REVOKED_SIGN.length(), origin.length() - Action.REVOKED_SIGN.length());
+        } else {
             //若不是，则将该操作加入栈中，待恢复时取用
             int sourceLength = end - start;
             int changedLength = dend - dstart;
-            if(changedLength == 0 && sourceLength > 0){
+            if (changedLength == 0 && sourceLength > 0) {
                 //此时新增输入，可能是从键盘，也可能是直接粘贴
-                int addEnd = source.equals("\n")?dstart+5 : dstart+sourceLength;
-                mBackStack.push(new AddAction(dstart,addEnd));
-            }else if(sourceLength == 0 && changedLength>0 ){
+                int addEnd = source.equals("\n") ? dstart + 5 : dstart + sourceLength;
+                mBackStack.push(new AddAction(dstart, addEnd));
+            } else if (sourceLength == 0 && changedLength > 0) {
                 //此时为删除
-                mBackStack.push(new DeleteAction(dest.toString().substring(dstart,dend),dstart));
-            }else if(sourceLength > 0 && changedLength > 0){
+                mBackStack.push(new DeleteAction(dest.toString().substring(dstart, dend), dstart));
+            } else if (sourceLength > 0 && changedLength > 0) {
                 //此时为替换粘贴
-                mBackStack.push(new ReplaceAction(dest.toString().substring(dstart,dend),dstart,sourceLength));
+                mBackStack.push(new ReplaceAction(dest.toString().substring(dstart, dend), dstart, sourceLength));
             }
-            if(mBackStack.size()>= MAX_ITEM_IN_STACK)
+            if (mBackStack.size() >= MAX_ITEM_IN_STACK) {
                 compressBackStack();//控制栈的长度
+            }
             return source;
         }
 
@@ -53,12 +60,14 @@ public class RevokeInputFilter implements InputFilter {
     /**
      * 清除栈的数据，使之始终保存在阈值之下
      */
-    private void compressBackStack(){
+    private void compressBackStack() {
         Stack<Action> tempStack = new Stack<>();
-        for(int i = 0;i < MAX_ITEM_IN_STACK/2; i++)
+        for (int i = 0; i < MAX_ITEM_IN_STACK / 10; i++) {
             tempStack.push(mBackStack.pop());
+        }
         mBackStack.clear();
-        for(int i = 0;i < tempStack.size(); i++)
+        for (int i = 0; i < tempStack.size(); i++) {
             mBackStack.push(tempStack.pop());
+        }
     }
 }
